# backend/openapi.yaml
openapi: 3.1.0
...
info:
  title: LoopBuy Backend API
  version: 1.0.0
  description: >-
    REST API for LoopBuy accounts, marketplace listings, social features,
    scam screening, recommendations, and the RAG shopping assistant.
servers:
  - url: http://localhost:8090
    description: Docker Compose local API
security: []
...
paths:
...
  /api/v1/auth/email/verify:
    post:
      tags: [Auth]
      operationId: verifyEmail
      summary: Consume a single-use email verification token
      description: POST is required so link scanners cannot verify an account merely by opening the email URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '204':
          description: Email verified. The token and every superseded token for the account are no longer usable.
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      summary: Revoke every refresh token for the current user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All refresh tokens revoked.
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token issued by register, login, or refresh.
...
  responses:
    Problem:
      description: Error response. Typical statuses are 400, 401, 403, 404, 409, 422, 428, 500, or 503.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    EmailVerificationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 90
          maxLength: 90
          description: Versioned HMAC-signed opaque token; stored only as a SHA-256 hash.
          writeOnly: true
...